refactor(common): extract mapper-agnostic merge logic from RequestMerger
